refactor: 코드 마이너 리팩토링
