+++ client/src/components/RoomSettingModal/SelectProblem.tsx
@@ -19,7 +19,6 @@ export default function SelectProblem({
   const onChangeInput = async (event: React.ChangeEvent<HTMLInputElement>) => {
     setSearchInput(event.target.value);
 
-    // if blank or whitespace, return
     if (!event.target.value.trim()) return;
   };
 
@@ -31,12 +30,11 @@ export default function SelectProblem({
     const newProblem: ProblemType = {
       title: option.title,
       boj_problem_id: option.bojProblemId,
-      url: `https://www.acmicpc.net/problem/${option.bojProblemId}}`,
+      url: `https://www.acmicpc.net/problem/${option.bojProblemId}`,
       level: option.level,
-      // TODO : response have no tag attribute
+      // response have no tag attribute, leave blank
       tag: [],
     };
-    // max 4 problems
     if (problemList.length + 1 > 4) {
       alert('문제는 최대 4개까지만 등록할 수 있습니다.');
       return;
