+++ client/src/components/buttons/StartButton.tsx
@@ -14,7 +14,7 @@ export default function StartButton({ isHost }: { isHost: boolean }) {
     <div className="flex w-full justify-end">
       <button
         onClick={start}
-        className="bg-accent text-default_white flex h-[33px] w-full items-center justify-center rounded-lg font-medium hover:opacity-80">
+        className="flex h-[33px] w-full items-center justify-center rounded-lg bg-accent font-medium text-default_white hover:opacity-80">
         {timer}
       </button>
     </div>
