+++ client/src/contexts/AuthProvider.tsx
@@ -1,21 +1,15 @@
-import { useNavigate } from 'react-router-dom';
-import {
-  ReactNode,
-  createContext,
-  useContext,
-  useEffect,
-  useState,
-} from 'react';
+import { useNavigate, useLocation } from 'react-router-dom';
+import { ReactNode, createContext, useEffect, useState } from 'react';
 import { CreateUser } from '../types/CreateUserType';
 import { logout } from '../apis/logout';
 import { getSession } from '../apis/getSession';
 import { useLocalStorage } from './LocalStorageProvider';
 
-interface AuthContextType {
+export interface AuthContextType {
   user: CreateUser | null;
 }
 
-interface AuthUpdateType {
+export interface AuthUpdateType {
   onLogout: () => void;
 }
 
@@ -23,11 +17,16 @@ interface AuthProviderProps {
   children: ReactNode;
 }
 
-const AuthContext = createContext<AuthContextType>({} as AuthContextType);
-const AuthUpdateContext = createContext<AuthUpdateType>({} as AuthUpdateType);
+export const AuthContext = createContext<AuthContextType>(
+  {} as AuthContextType,
+);
+export const AuthUpdateContext = createContext<AuthUpdateType>(
+  {} as AuthUpdateType,
+);
 
 export const AuthProvider: React.FC<AuthProviderProps> = ({ children }) => {
   const navigate = useNavigate();
+  const location = useLocation();
   const { removeItem } = useLocalStorage();
 
   const [user, setUser] = useState<CreateUser | null>(null);
@@ -45,7 +44,7 @@ export const AuthProvider: React.FC<AuthProviderProps> = ({ children }) => {
         setUser(data);
       }
     });
-  }, []);
+  }, [location]);
 
   return (
     <AuthContext.Provider value={{ user }}>
@@ -54,20 +53,4 @@ export const AuthProvider: React.FC<AuthProviderProps> = ({ children }) => {
       </AuthUpdateContext.Provider>
     </AuthContext.Provider>
   );
-};
-
-export const useAuthContext = (): AuthContextType => {
-  const context = useContext(AuthContext);
-  if (!context) {
-    throw new Error('useAuthContext must be used within a AuthProvider');
-  }
-  return context;
-};
-
-export const useAuthUpdateContext = (): AuthUpdateType => {
-  const context = useContext(AuthUpdateContext);
-  if (!context) {
-    throw new Error('useAuthUpdateContext must be used within a AuthProvider');
-  }
-  return context;
 };
