+++ client/src/contexts/LocalStorageProvider.tsx
@@ -1,4 +1,3 @@
-// localStorageContext.tsx
 import React, { createContext, useContext, ReactNode } from 'react';
 
 interface LocalStorageContextProps {
