+++ client/src/contexts/useTheme.tsx
@@ -1,0 +1,10 @@
+import { useContext } from 'react';
+import { ThemeContextType, ThemeContext } from './ThemeProvider';
+
+export const useTheme = (): ThemeContextType => {
+  const context = useContext(ThemeContext);
+  if (!context) {
+    throw new Error('useTheme must be used within a ThemeProvider');
+  }
+  return context;
+};
