+++ client/src/pages/NotFound.tsx
@@ -1,7 +1,7 @@
 export default function NotFound() {
   return (
-    <>
-      <h1>Not Found!</h1>
-    </>
+    <div className="flex min-h-screen items-center justify-center">
+      <h1 className="text-center text-5xl">Not Found!</h1>
+    </div>
   );
 }
