+++ client/src/util/goSolveProblem.ts
@@ -1,0 +1,5 @@
+import { ProblemType } from "../types/ProblemType";
+
+export const goSolveProblem = (problem:ProblemType) => () => {
+    window.open(problem.url);
+  };
